Moved classes to separate files, Item delete function bug fixing
