+++ frontend/GW_UI/GW_UI/Classes/TypeItem.cs
@@ -1,0 +1,13 @@
+﻿using System.Text.Json.Serialization;
+
+namespace GW_UI
+{
+    public class TypeItem
+    {
+        [JsonPropertyName("id")]
+        public int ID { get; set; }
+
+        [JsonPropertyName("full_name")]
+        public string TypeName { get; set; }
+    }
+}
